🔨 Refactor code
